Bugs Fixed && Logger Addition on Production
Size Buffer on ServerComm has increased to 4096 because the server_param. Maybe is better to just throw this on trash and mantain 512.

Correction of type in TacticalFormations

All the bugs are verified. I am just afraid that the positions on Default Formation may no be accepted by the C++ code on the same order.

+++ src/booting/TacticalFormations.hpp
@@ -2,7 +2,7 @@
 
 // Deve seguir exatamente o padrão que colocamos
 namespace TacticalFormations {
-	inline constexpr int Default[] = {
+	inline constexpr double Default[] = {
 		-33.17141941169382, -25.62993762994164,
 		-33.17141941169382, -12.823284823286832,
 		-33.17141941169382, -0.01663201663202301,
